Table & Short Views (#4)
* Customized views
- added the short and table views.
- Added option while configuring to have yes, no, table or short views

* Customized views
- test for table in sysinfo

* View changes

+++ remove.sh
@@ -4,35 +4,37 @@
 MOTD_DIR="/etc/update-motd.d"
 BACKUP_DIR="/etc/update-motd.d.bak"
 
-# Check for root privileges
-if [ "$EUID" -ne 0 ]; then
-  echo "Error: Please run as root (sudo ./uninstall.sh)"
-  exit 1
-fi
+# Load Common Scripts
+source ./common.sh
 
-echo "Uninstalling Linux MOTD..."
+draw_line
+echo -e "${LCYAN}Uninstalling Linux MOTD...${NC}"
+draw_line
+
+validate_root_privileges
+get_config_script
 
 # 1. Remove current files
-echo "Removing installed files from $MOTD_DIR..."
-rm -rf "$MOTD_DIR"/*
+echo -e "${LGREEN}Removing installed files from $MOTD_DIR...${NC}"
+rm -rf "${MOTD_DIR:?}/"*
 
 # 2. Restore backup
 if [ -d "$BACKUP_DIR" ]; then
-    echo "Restoring original files from $BACKUP_DIR..."
+    echo -e "${LGREEN}Restoring original files from $BACKUP_DIR...${NC}"
     if [ "$(ls -A $BACKUP_DIR)" ]; then
         cp -r "$BACKUP_DIR/"* "$MOTD_DIR/"
         chmod +x "$MOTD_DIR"/*
-        echo "Backup restored successfully."
+        echo -e "${GREEN}Backup restored successfully.${NC}"
     else
-        echo "Warning: Backup directory exists but is empty."
+        echo -e "${RED}Warning: Backup directory exists but is empty.${NC}"
     fi
     
     # Optional: Remove backup directory after restore?
     # rm -rf "$BACKUP_DIR"
-    echo "Note: Backup folder $BACKUP_DIR was kept."
+    echo -e "${CYAN}Note:${NC} Backup folder $BACKUP_DIR was kept."
 else
-    echo "Warning: No backup directory found at $BACKUP_DIR. Directory $MOTD_DIR is now empty."
+    echo -e "${RED}Warning: No backup directory found at $BACKUP_DIR. Directory $MOTD_DIR is now empty.${NC}"
 fi
 
-echo "------------------------------------------------"
-echo "Uninstallation complete."
+draw_line
+echo -e "${GREEN}Uninstallation complete.${NC}\n"
